minimum version of final project api success fully done

+++ app/final_project_api/views/business/business_views.py
@@ -41,18 +41,16 @@ blp = Blueprint(
     "business",
     __name__,
     url_prefix="/api/business",
-    description="""
-                business management end point
-                """,
+    description="business rout use to manage user business data ",
 )
 
 
-@blp.route("/")
+@blp.route("")
 class BusinessViews(MethodView):
     @blp.arguments(schema=QuerySchema, location="query")
     @blp.response(schema=BusinessPublicSchema(many=True), status_code=HTTPStatus.OK)
     def get(self, query_data: QueryData):
-
+        """As a user, i can get all business public data without auth"""
         return BusinessModel.get_all_public_model(query_data=query_data)
 
     @jwt_required()
@@ -63,8 +61,10 @@ class BusinessViews(MethodView):
         description=MessageService.get_message("failed_to_add_business").format(
             "pass on payload"
         ),
+        example=MessageService.get_message("failed_to_create_business_example"),
     )
     def post(self, business_data: BusinessCreateData):
+        """As a auth user, i can create my business"""
         try:
             model = BusinessModel.add_model(
                 user_id=getCurrentAuthId(),
@@ -95,6 +95,7 @@ class BusinessByIdViews(MethodView):
         ),
     )
     def get(self, business_id: str):
+        """As a user i can get business detail by it's id"""
         businessModel: BusinessModel = BusinessModel.get_model_by_id(
             model_id=business_id
         )
@@ -120,12 +121,30 @@ class BusinessByIdViews(MethodView):
             "pass on url "
         ),
     )
+    @blp.alt_response(
+        status_code=HTTPStatus.FORBIDDEN,
+        description=MessageService.get_message("failed_to_update_business").format(
+            "on url, it might be because user id on business not match with auth user id"
+        ),
+    )
     def put(self, business_data: BusinessCreateData, business_id: str):
+        """As a user i can update my own business data"""
+        businessModel: BusinessModel = BusinessModel.get_model_by_id(
+            model_id=business_id
+        )
+        if businessModel.user_id != getCurrentAuthId():
+            abort(
+                http_status_code=HTTPStatus.FORBIDDEN,
+                message=MessageService.get_message("failed_to_delete_business").format(
+                    f"{business_id}, business user id not match with auth user id"
+                ),
+            )
         try:
             return BusinessModel.update_by_id(
                 business_id=business_id, update_data=business_data
             )
         except Exception as e:
+            print(e)
             abort(
                 http_status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
                 message=MessageService.get_message("failed_to_update_business").format(
@@ -140,10 +159,27 @@ class BusinessByIdViews(MethodView):
             "pass on url"
         ),
     )
+    @blp.alt_response(
+        status_code=HTTPStatus.FORBIDDEN,
+        description=MessageService.get_message("failed_to_update_business").format(
+            "on url, it might be because user id on business not match with auth user id"
+        ),
+    )
     def delete(self, business_id: str):
+        """As a user i, can delete my own business"""
+        businessModel: BusinessModel = BusinessModel.get_model_by_id(
+            model_id=business_id
+        )
+        if businessModel.user_id != getCurrentAuthId():
+            abort(
+                http_status_code=HTTPStatus.FORBIDDEN,
+                message=MessageService.get_message("failed_to_delete_business").format(
+                    f"{business_id}, business user id not match with auth user id"
+                ),
+            )
         try:
             BusinessModel.delete_by_id(model_id=business_id)
-            return {"message": f"business with id {business_id} success filly delete"}
+            return {"message": f"business with id {business_id} success fully delete"}
         except Exception as e:
             abort(
                 http_status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
@@ -154,7 +190,7 @@ class BusinessByIdViews(MethodView):
 
 
 # TODO: handle product  image
-@blp.route("/image/<string:product_id>")
+@blp.route("/image/<string:business_id>")
 class ImageUserViews(MethodView):
 
     @jwt_required()
@@ -164,7 +200,8 @@ class ImageUserViews(MethodView):
         status_code=HTTPStatus.UNPROCESSABLE_ENTITY,
         description=MessageService.get_message("file_not_image_type"),
     )
-    def post(self, data: ImagesPayloadData, product_id: str):
+    def post(self, data: ImagesPayloadData, business_id: str):
+        """As a user, i can post business image"""
         image_data: Upload = data.image
         if ImageService.check_extension(image_data=image_data) != True:
             abort(
@@ -176,7 +213,7 @@ class ImageUserViews(MethodView):
             userImageModel = BusinessImageModel.add_model(
                 public_id=response_data.public_id,
                 secure_url=response_data.secure_url,
-                product_id=product_id,
+                business_id=business_id,
             )
             return userImageModel
         except Exception as e:
@@ -191,19 +228,14 @@ class ImageUserViews(MethodView):
             "from jwt token"
         ),
     )
-    def get(self, product_id: str):
-        """_summary_
-        Args:
-            product_id (str): _description_
-        Returns:
-            _type_: _description_
-        """
-        userModel: BusinessModel = BusinessModel.get_model_by_id(model_id=product_id)
+    def get(self, business_id: str):
+        """As a user, i can see all of my image"""
+        userModel: BusinessModel = BusinessModel.get_model_by_id(model_id=business_id)
         if bool(userModel) == False:
             abort(
                 http_status_code=HTTPStatus.CONFLICT,
                 message=MessageService.get_message("product_not_found").format(
-                    product_id
+                    business_id
                 ),
             )
         return userModel
@@ -219,8 +251,10 @@ class ImageUserViews(MethodView):
         description=MessageService.get_message("image_not_found").format(
             "pass on url string"
         ),
+        example=MessageService.get_message("image_not_found_example"),
     )
-    def get(self, product_id: str, image_id: str):
+    def get(self, image_id: str):
+        """As a user, i can see my image by its id"""
         imageMode: BusinessImageModel = BusinessImageModel.get_model_by_id(
             model_id=image_id
         )
@@ -238,8 +272,10 @@ class ImageUserViews(MethodView):
         description=MessageService.get_message("image_not_found").format(
             "pass on url string"
         ),
+        example=MessageService.get_message("image_not_found_example"),
     )
-    def delete(self, product_id: str, image_id: str):
+    def delete(self, image_id: str):
+        """As a user, i can delete my business image"""
         imageMode: BusinessImageModel = BusinessImageModel.get_model_by_id(
             model_id=image_id
         )
@@ -259,6 +295,7 @@ class ImageUserViews(MethodView):
         description=MessageService.get_message("image_not_found").format(
             "pass on url string"
         ),
+        example=MessageService.get_message("image_not_found_example"),
     )
     @blp.alt_response(
         status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
@@ -266,7 +303,8 @@ class ImageUserViews(MethodView):
             "pass on the url string"
         ),
     )
-    def put(self, product_id: str, image_id: str):
+    def put(self, image_id: str):
+        """as a user, i can update my business image"""
         imageMode: BusinessImageModel = BusinessImageModel.get_model_by_id(
             model_id=image_id
         )
